セルのコンポーネントに親子関係をつけた
